requirements.txt
pip>=24.0
setuptools>=69.0.3
wheel>=0.42.0
streamlit==1.31.1
numpy==1.26.4
pandas==2.2.0
plotly==5.18.0
scikit-learn==1.4.0
umap-learn==0.5.5
hdbscan==0.8.33
python-dotenv==1.0.1
openai>=1.0.0
anthropic>=0.18.0
matplotlib>=3.8.0
